Move history refresh to base class
